+++ Changelog.md
@@ -12,6 +12,7 @@ Language Features:
 
 
 Compiler Features:
+ * Implement packed encoding for ABIEncoderV2.
  * C API (``libsolc`` / raw ``soljson.js``): Introduce ``solidity_free`` method which releases all internal buffers to save memory.
  * Commandline interface: Adds new option ``--new-reporter`` for improved diagnostics formatting
    along with ``--color`` and ``--no-color`` for colorized output to be forced (or explicitly disabled).